requirements.txt
# Core Framework Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0  # : отдельный пакет для BaseSettings

# Dependency Injection
dependency-injector==4.41.0

# Logging & Monitoring
structlog==23.2.0
loguru==0.7.2

# Discord & Telegram
aiohttp==3.9.1
requests==2.31.0
pyTelegramBotAPI==4.14.0
websockets==12.0

# Configuration & Environment
python-dotenv==1.0.0

# Data Processing
orjson==3.9.10
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Production Dependencies
redis==5.0.1
prometheus-client==0.19.0
psutil==5.9.6

# Development Dependencies (optional)
black==23.11.0
isort==5.12.0
mypy==1.7.0

# Additional utilities for token testing
colorama==0.4.6  # For colored terminal output